Proj1: another working lcl version

+++ Proj1/schedule.c
@@ -17,6 +17,9 @@ typedef struct
   pid_t pid;
 } Process;
 
+int ALRM = 0;
+int CHLD = 0;
+
 void print_processes(Process processes[], int n_processes)
 {
   printf("Number of Processes: %d\n", n_processes);
@@ -44,6 +47,22 @@ void print_processes(Process processes[], int n_processes)
 
 void signal_handler(int signum, siginfo_t *si, void *unused)
 {
+  if (signum == SIGALRM)
+  {
+    /*
+    printf("in signal_hanlder SIGALRM\n");
+    fflush(stdout);
+    */
+    ALRM = 1;
+  }
+  else if (signum == SIGCHLD)
+  {
+    /*
+    printf("in signal_hanlder SIGCHLD\n");
+    fflush(stdout);
+    */
+    CHLD = 1;
+  }
 }
 
 int main(int argc, char *argv[])
@@ -125,8 +144,10 @@ int main(int argc, char *argv[])
     n_processes++;
   }
 
+  /*
   print_processes(processes, n_processes);
   printf("done printing processes\n\n");
+  */
 
   sigset_t mask;
   struct itimerval timer;
@@ -165,21 +186,39 @@ int main(int argc, char *argv[])
       if (processes[i].pid > 0)
       {
         kill(processes[i].pid, SIGCONT);
+        ALRM = 0;
+        CHLD = 0;
 
-        usleep(quantum * 1000);
-        kill(processes[i].pid, SIGSTOP);
-
-        pid_t pid;
-        while ((pid = waitpid(processes[i].pid, NULL, WNOHANG)) > 0)
+        while (1)
         {
-          live_processes--;
-          for (j = 0; j < n_processes; j++)
+
+          if (ALRM)
           {
-            if (processes[j].pid == pid)
+            kill(processes[i].pid, SIGSTOP);
+            break;
+          }
+          else if (CHLD)
+          {
+            pid_t pid;
+            while ((pid = waitpid(processes[i].pid, NULL, WNOHANG)) > 0)
             {
-              processes[j].pid = -1;
-              break;
+              live_processes--;
+              for (j = 0; j < n_processes; j++)
+              {
+                if (processes[j].pid == pid)
+                {
+                  processes[j].pid = -1;
+                  break;
+                }
+              }
             }
+
+            if (setitimer(ITIMER_REAL, &timer, NULL) == -1)
+            {
+              perror("setitimer");
+              exit(EXIT_FAILURE);
+            }
+            break;
           }
         }
       }
